+++ skoolkit/config.py
@@ -140,6 +140,7 @@ COMMANDS = {
         'DefaultSnapshotFormat': ('z80', ''),
         'ScreenScale': (2, ''),
         'ScreenFps': (50, ''),
+        'TraceHeader': ('', ''),
         'TraceLine': ('${pc:04X} {i}', ''),
         'TraceOperand': ('$,02X,04X', ''),
         'UserAgent': ('', 'user_agent'),